Fixed GNU warning
git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/ptlib/trunk@27593 023b2edf-31b2-4de3-b41e-bca80c47788f

+++ src/ptclib/vartype.cxx
@@ -518,7 +518,7 @@ unsigned PVarType::AsUnsigned() const
         return UINT_MAX;
       return (unsigned)m_.floatExtended;
     case VarTime :
-      if (m_.time.seconds > UINT_MAX)
+      if ((unsigned)m_.time.seconds > UINT_MAX)
         return UINT_MAX;
       return (unsigned)m_.time.seconds;
     case VarGUID :
